requirements.txt
aiohttp>=3.9.0
aiofiles>=23.2.0
polars>=0.20.0
pyarrow>=14.0.0
numpy>=1.24.0
scipy>=1.11.0
requests>=2.31.0
python-dotenv>=1.0.0
cupy-cuda12x>=12.0.0